Reuse safetensors validation and consolidate checkpoint adapters
