+++ experimental/lite/megatron/lite/model/deepseek_v41/lite/checkpoint.py
@@ -1,9 +1,181 @@
 # Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-"""Numerical bindings of V4.1 checkpoint bytes; archival storage stays unchanged."""
+"""Immutable file-backed checkpoint bytes, independent of model execution."""
 
+import hashlib
+import json
+import os
+import struct
+import tempfile
+from dataclasses import asdict, dataclass
+from pathlib import Path
+from types import MappingProxyType
+
+from safetensors import SafetensorError, safe_open
 import torch
 from megatron.lite.primitive.quantization.block_fp8 import dequantize_block_fp8
 from megatron.lite.primitive.quantization.mxfp4 import dequantize_mxfp4
+
+_CHUNK = 8 * 1024 * 1024
+
+
+def validate_execution(*, enable_dspark_execution: bool = False) -> None:
+    if type(enable_dspark_execution) is not bool:
+        raise TypeError("enable_dspark_execution must be bool")
+    if enable_dspark_execution:
+        raise NotImplementedError("DSpark execution is not implemented")
+
+
+def _unique_object(pairs):
+    result = {}
+    for key, value in pairs:
+        if key in result:
+            raise ValueError(f"duplicate header key: {key}")
+        result[key] = value
+    return result
+
+
+def _coverage(entries, expected_keys):
+    expected = list(expected_keys)
+    if len(expected) != len(set(expected)):
+        raise ValueError("duplicate expected keys")
+    missing, extra = set(expected) - entries.keys(), entries.keys() - set(expected)
+    if missing or extra:
+        raise ValueError(
+            f"key coverage mismatch: missing={len(missing)} extra={len(extra)}"
+        )
+
+
+@dataclass(frozen=True)
+class TensorEntry:
+    release_key: str
+    dtype: str
+    shape: tuple[int, ...]
+    byte_length: int
+    source_shard: str
+    offset: int
+    payload_digest: str
+
+
+def _stream(entry, output=None):
+    digest = hashlib.sha256()
+    with open(entry.source_shard, "rb") as source:
+        source.seek(entry.offset)
+        remaining = entry.byte_length
+        while remaining:
+            raw = source.read(min(remaining, _CHUNK))
+            if not raw:
+                raise ValueError(f"truncated payload: {entry.release_key}")
+            remaining -= len(raw)
+            digest.update(raw)
+            if output is not None:
+                output.write(raw)
+    return digest.hexdigest()
+
+
+class CheckpointTensorStore:
+    """Byte ranges with validated metadata; no tensor/model allocation or casting."""
+
+    def __init__(self, entries):
+        self.entries = MappingProxyType(dict(sorted(entries.items())))
+
+    @classmethod
+    def load(cls, paths, *, expected_keys, key_prefix=None):
+        entries = {}
+        for path in paths:
+            path = Path(path).resolve()
+            # Delegate shape, dtype, offset and complete-file validation to the
+            # safetensors reader; retain byte ranges for lossless archival copies.
+            try:
+                with safe_open(path, framework="pt") as source:
+                    keys = list(source.keys())
+            except SafetensorError as error:
+                raise ValueError(f"invalid safetensors file: {path}") from error
+            with path.open("rb") as source:
+                length = struct.unpack("<Q", source.read(8))[0]
+                header = json.loads(source.read(length), object_pairs_hook=_unique_object)
+            for name in keys:
+                if key_prefix is not None and not name.startswith(key_prefix):
+                    continue
+                if name in entries:
+                    raise ValueError(f"duplicate tensor: {name}")
+                record = header[name]
+                start, end = record["data_offsets"]
+                entry = TensorEntry(
+                    name, record["dtype"], tuple(record["shape"]),
+                    end - start, str(path), 8 + length + start, "",
+                )
+                entries[name] = TensorEntry(
+                    **{**asdict(entry), "payload_digest": _stream(entry)}
+                )
+        _coverage(entries, expected_keys)
+        return cls(entries)
+
+    def manifest(self):
+        return {name: asdict(entry) for name, entry in self.entries.items()}
+
+    def read(self, name):
+        import io
+
+        output = io.BytesIO()
+        entry = self.entries[name]
+        if _stream(entry, output) != entry.payload_digest:
+            raise ValueError(f"payload digest mismatch: {name}")
+        return output.getvalue()
+
+    def shard(self, rank, world_size):
+        if (
+            type(world_size) is not int
+            or world_size < 1
+            or type(rank) is not int
+            or not 0 <= rank < world_size
+        ):
+            raise ValueError("invalid archival rank/world size")
+        return type(self)(
+            {key: self.entries[key] for key in list(self.entries)[rank::world_size]}
+        )
+
+    @classmethod
+    def merge(cls, stores, *, expected_keys):
+        entries = {}
+        for store in stores:
+            for name, entry in store.entries.items():
+                if name in entries:
+                    raise ValueError(f"duplicate tensor: {name}")
+                entries[name] = entry
+        _coverage(entries, expected_keys)
+        return cls(entries)
+
+    def save(self, path):
+        """Stream to a new file, publishing only after every payload digest agrees."""
+        path = Path(path)
+        if path.exists():
+            raise FileExistsError(path)
+        header, cursor = {}, 0
+        for name, entry in self.entries.items():
+            header[name] = dict(
+                dtype=entry.dtype,
+                shape=list(entry.shape),
+                data_offsets=[cursor, cursor + entry.byte_length],
+            )
+            cursor += entry.byte_length
+        encoded = json.dumps(header, separators=(",", ":")).encode()
+        encoded += b" " * (-len(encoded) % 8)
+        fd, temporary = tempfile.mkstemp(prefix=".checkpoint-", dir=path.parent)
+        try:
+            with os.fdopen(fd, "wb") as output:
+                output.write(struct.pack("<Q", len(encoded)))
+                output.write(encoded)
+                for name, entry in self.entries.items():
+                    if _stream(entry, output) != entry.payload_digest:
+                        raise ValueError(f"payload digest mismatch: {name}")
+                output.flush()
+                os.fsync(output.fileno())
+            # Exclusive publication also prevents a concurrent writer from being overwritten.
+            os.link(temporary, path)
+        finally:
+            os.unlink(temporary)
+
+
 
 _TORCH_DTYPES = {
     "I8": torch.int8,
@@ -199,7 +371,6 @@ def save_model(model, path):
     import tempfile
     from pathlib import Path
 
-    from .checkpoint_store import CheckpointTensorStore, TensorEntry
 
     path = Path(path)
     if path.exists():
@@ -249,7 +420,6 @@ def load_model(model, path):
     import json
     from pathlib import Path
 
-    from .checkpoint_store import CheckpointTensorStore
     from .engram import EngramTable
 
     path = Path(path)
